refactor: sdk 미니 플레이어 컴포넌트 분리
